Revise meteorological station's location error

Revise meteorological station's location error

revise_locaiton(
  d,
  prefix = "",
  dist_max = 100,
  score_min = 2,
  vars_rm = c("date_begin", "date_end", "n_all"),
  verbose = 2
)

revise_locaiton_multi(info, dist_max = 50)

Arguments

d

A data.frame with the columns at least of `dist`, `lon`, `lat`, `alt` and `n_period`, `QC`

dist_max

If the site moving distance beyond the `dist_max`, it will be regarded as outlier and will be fixed by [revise_location()].

score_min

if distance score lower than `score_min`, will not fix

vars_rm

variables not shown in the console

info

site moving info returned by [get_moveInfo()]

Value

- sites with not outliers, QC = ""; - sites with outliers, for outliers, QC equals: score = 0: unfixed score = 1: marginal quality fixed score = 2: good quality fixed

Details

Continuous moving location outliers are also considered at here.

Examples

d = st_moveInfo[site == 50772]
revise_locaiton_multi(d)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> [01th:50772] =======================================
#>     [tag=1] ------------------------------------------- 
#>     site moveTimes tag   lon  lat   alt period_date_begin period_date_end
#> 1: 50772         6   1 12951 4752 283.5        1957-10-01      1962-12-31
#> 2: 50772         6   2 12951 4752 293.5        1963-01-01      1964-12-31
#> 3: 50772         6   3 12951 4752 296.5        1965-01-01      1970-12-31
#> 4: 50772         6   4 12915 4807 296.5        1971-01-01      2003-12-31
#>    n_period  dist QC
#> 1:     1918 69.11   
#> 2:      731 69.11   
#> 3:     2191 69.11   
#> 4:    12053  0.00   
#> ---
#> [w]: tag=1 unfixed, low location matched score (0)
#>     [tag=2] ------------------------------------------- 
#>     site moveTimes tag   lon  lat   alt period_date_begin period_date_end
#> 1: 50772         6   1 12951 4752 283.5        1957-10-01      1962-12-31
#> 2: 50772         6   2 12951 4752 293.5        1963-01-01      1964-12-31
#> 3: 50772         6   3 12951 4752 296.5        1965-01-01      1970-12-31
#> 4: 50772         6   4 12915 4807 296.5        1971-01-01      2003-12-31
#>    n_period  dist      QC
#> 1:     1918 69.11 unfixed
#> 2:      731 69.11        
#> 3:     2191 69.11        
#> 4:    12053  0.00        
#> ---
#> [w]: tag=2 unfixed, low location matched score (0)
#>     [tag=3] ------------------------------------------- 
#>     site moveTimes tag   lon  lat   alt period_date_begin period_date_end
#> 1: 50772         6   1 12951 4752 283.5        1957-10-01      1962-12-31
#> 2: 50772         6   2 12951 4752 293.5        1963-01-01      1964-12-31
#> 3: 50772         6   3 12951 4752 296.5        1965-01-01      1970-12-31
#> 4: 50772         6   4 12915 4807 296.5        1971-01-01      2003-12-31
#>    n_period  dist      QC
#> 1:     1918 69.11 unfixed
#> 2:      731 69.11 unfixed
#> 3:     2191 69.11        
#> 4:    12053  0.00        
#> ---
#> [w]: tag=3 suspicious, low location matched score (1)
#> [info] 0 sites fixed, 0 sites unfixed, 1 sites not sure
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>     site moveTimes tag   lon  lat   alt period_date_begin period_date_end
#> 1: 50772         6   1 12951 4752 283.5        1957-10-01      1962-12-31
#> 2: 50772         6   2 12951 4752 293.5        1963-01-01      1964-12-31
#> 3: 50772         6   3 12951 4752 296.5        1965-01-01      1970-12-31
#> 4: 50772         6   4 12915 4807 296.5        1971-01-01      2003-12-31
#> 5: 50772         6   5 12914 4806 288.0        2004-01-01      2017-12-31
#> 6: 50772         6   6 12914 4807 300.0        2018-01-01      2018-12-31
#>    date_begin   date_end n_all n_period  dist
#> 1: 1957-10-01 2018-12-31 22372     1918 69.11
#> 2: 1957-10-01 2018-12-31 22372      731 69.11
#> 3: 1957-10-01 2018-12-31 22372     2191 69.11
#> 4: 1957-10-01 2018-12-31 22372    12053  0.00
#> 5: 1957-10-01 2018-12-31 22372     5114  2.20
#> 6: 1957-10-01 2018-12-31 22372      365  1.86